Fix reading DVD zip files (GC mode).  Strange Snes9x didn't need this.  Also, the DVD fixes have been pulled from Crunchy2's Snes9xGX v2.0.1b8.

+++ source/drivers/gamecube/gcdvd.c
@@ -710,7 +710,7 @@ int LoadDVDFile( unsigned char *buffer )
             memcpy(&buffer[offset], &readbuffer, i);
         }
     } else {
-        return unzipDVDFile( buffer, discoffset, rootdirlength);
+        return unzipDVDFile( buffer, (u32)discoffset, rootdirlength);
     }
     if (UseSDCARD) SDCARD_CloseFile (filehandle);
